+++ src/app/lists/urls_brands.js
@@ -1,0 +1,5 @@
+export default [
+    'https://www.consumerreports.org/cars/kia/',
+    'https://www.consumerreports.org/cars/honda/',
+    'https://www.consumerreports.org/cars/subary/',
+];