+++ src/app/lists/urls_default.js
@@ -1,0 +1,5 @@
+export default [
+    'https://www.consumerreports.org/cars/kia/', // 200 OK
+    'http://cr.org/burgers', // 404 with redirect
+    'bad.url', // no good
+];