+++ src/app/lists/urls_models.js
@@ -1,0 +1,5 @@
+export default [
+    'https://www.consumerreports.org/cars/kia/k5/',
+    'https://www.consumerreports.org/cars/honda/accord/',
+    'https://www.consumerreports.org/cars/subary/burgers',
+];